requirements.txt
docker-compose
pykube
google-python-api-client
click